requirements.txt
crochet==2.1.1
deep-translator==1.11.4
Flask==3.0.3
Flask-Session==0.8.0
itemadapter==0.9.0
mysql-connector-python==9.0.0
python_dateutil==2.9.0.post0
Requests==2.32.3
Scrapy==2.11.2
Twisted==24.7.0
Werkzeug==3.0.4
